scr/utils.py
from pdfminer.high_level import extract_text
import docx


def extract_text_from_pdf(pdf_path):
    return extract_text(pdf_path)


def extract_text_from_docx(docx_path):
    doc = docx.Document(docx_path)
    return "\n".join([para.text for para in doc.paragraphs])


class ModelError(Exception):
    pass
