requirements.txt
streamlit
streamlit-pills
langchain
langchain-community
pdfminer-six
docx
streamlit-aggrid==0.2.2-2
